+++ .gitignore
@@ -1,6 +1,7 @@
 # vim
 *.swp
 *.swo
+*.swm
 # Logs
 logs
 *.log
